refactor: rename websocket module to sse
Fold reactQueryHandler and the message schema into app/sse/ next to
SseConnection now that no WebSocket transport remains. Drop the
"websocket messages" mention from a few stale type doc-comments.
