+++ apps/web/src/app/sse/__tests__/schema.test.ts
