+++ apps/web/src/app/sse/schema.ts
